+++ partial-recall.el
@@ -941,8 +941,7 @@ are not considered."
     (define-key map (kbd "f") 'partial-recall-forget)
     (define-key map (kbd "i") 'partial-recall-implant)
     (define-key map (kbd "l") 'partial-recall-menu)
-    (define-key map (kbd "m") 'partial-recall-remember)
-    (define-key map (kbd "r") 'partial-recall-reinforce)
+    (define-key map (kbd "r") 'partial-recall-remember)
     (define-key map (kbd "s") 'partial-recall-lift)
     map)
   "Map for `partial-recall-mode' commands.")
@@ -975,13 +974,6 @@ all buffers."
   (interactive (list (partial-recall--complete-reality "Switch to moment: " t)))
 
   (switch-to-buffer buffer))
-
-;;;###autoload
-(defun partial-recall-reinforce (buffer)
-  "Reinforce BUFFER."
-  (interactive (list (partial-recall--complete-reality "Re-inforce moment: ")))
-
-  (partial-recall--reinforce buffer))
 
 ;;;###autoload
 (defun partial-recall-reclaim (buffer)